pyproject.toml
[tool.black]
line-length = 79
include = '\.pyi?$'

[tool.isort]
multi_line_output = 3
include_trailing_comma = true

[tool.poetry]
name = "interop_clients"
version = "0.1.0"
description = "Alternative interop clients for AUVSI SUAS"
authors = ["NC State Aerial Robotics Club <aerialroboticsclub@ncsu.edu>"]

[tool.poetry.dependencies]
python = "^3.6"
click = "^6.7"
requests = "^2.22.0"
pyzmq = "^18.1.1"
GDAL = { version = "^3.0.4", optional = true }
pyproj = "^2.4.2"

[tool.poetry.dev-dependencies]
pytest = "^3.0"
black = "^19.3b0"
flake8 = "^3.7.9"
isort = "^4.3.21"
mypy = "^0.761"
flake8-bugbear = "^20.1.0"

[tool.poetry.scripts]
interop = "interop_clients.cli:main"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
